// src/permissions/enums/permission-name.enum.ts
export enum PermissionName {
  // Incident
  INCIDENT_MASTER = 'incident_master',
  INCIDENT_USER = 'incident_user',
  INCIDENT_SUBMITTER = 'incident_submitter',
  INCIDENT_VIEWER = 'incident_viewer',

  // Foundation
  Foundation_SupportGroup = 'foundation_support_group',
  Foundation_People = 'foundation_people',
  Foundation_Category = 'foundation_category',
}